+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - PostgreSQL: Support partial indices (bug #1048)
 - PostgreSQL: Support calling functions returning table (bug #1040)
 - PostgreSQL: Add NOT ILIKE operator (bug #1066)
+- Editor: Fix bit and enum search (bug #1062)
 - Designs: adminer.css with 'prefers-color-scheme: dark' doesn't disable dark mode
 - Plugins: Method bodyClass() to add &lt;body class>
 - Plugins: Allow setting dark mode in css()